backend/requirements.txt
django
djangorestframework
chromadb
sentence-transformers
rank_bm25
redis
requests